TG_BOT/NewBotAs/FOR_BOT/requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
attrs==25.1.0
auto-py-to-exe==2.46.0
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
Eel==0.18.1
frozenlist==1.5.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
MouseInfo==0.1.3
mss==10.0.0
multidict==6.1.0
packaging==24.2
pefile==2023.2.7
propcache==0.3.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
pytweening==1.2.0
pywin32==308
pywin32-ctypes==0.2.3
requests==2.32.3
screeninfo==0.8.1
setuptools==75.8.2
typing_extensions==4.12.2
urllib3==2.3.0
winshell==0.6
yarl==1.18.3
zope.event==5.0
